+++ app/scripts/storage/storage-gdrive.js
@@ -126,7 +126,7 @@ const StorageGDrive = StorageBase.extend({
         });
     },
 
-    list: function(callback) {
+    list: function(dir, callback) {
         this._oauthAuthorize((err) => {
             if (err) { return callback && callback(err); }
             this.logger.debug('List');