+++ release-notes.md
@@ -6,6 +6,7 @@ Release notes
 `*` prevent master password autocomplete  
 `*` build with node.js 8  
 `-` fixed tray icon click crash  
+`*` show usernames in in entry list  
 
 ##### v1.5.6 (2017-08-31)
 `-` fix #722: hang on start in desktop